Cache size is a long

+++ oak-blob-plugins/src/main/java/org/apache/jackrabbit/oak/plugins/blob/AbstractSharedCachingDataStore.java
@@ -91,8 +91,8 @@ public abstract class AbstractSharedCachingDataStore extends AbstractDataStore
     /**
      * Record cache size
      */
-    private static final int RECORD_CACHE_SIZE = SystemPropertySupplier
-            .create("oak.blob.recordcache.size", 10000)
+    private static final long RECORD_CACHE_SIZE = SystemPropertySupplier
+            .create("oak.blob.recordcache.size", 10000L)
             .loggingTo(LOG)
             .formatSetMessage( (name, value) -> String.format("%s set to: %s", name, value) )
             .get();
